refactor: move useThemeMode hook from components/ui/ into hooks/
